requirements.txt
alembic==1.16.4
annotated-types==0.7.0
apig-wsgi==2.19.0
Authlib==1.6.1
aws-lambda-wsgi==0.0.6
bcrypt==4.3.0
blinker==1.9.0
Bootstrap-Flask==2.5.0
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==3.3.2
cssselect2==0.8.0
defusedxml==0.7.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
Flask==3.1.1
Flask-Authlib==1.6.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.59.0
fpdf2==2.8.4
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psycopg2==2.9.10
pyasn1==0.6.1
pycountrycode==1.1.8
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
pytz==2025.2
reportlab==4.4.3
requests==2.32.4
rich==14.0.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.41
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
zopfli==0.2.3.post1
